backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # No necesario para SQLite
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Stripe
stripe==7.10.0

# WhatsApp & Communications
httpx==0.26.0
twilio==8.11.0
aiofiles==23.2.1

# Azure OpenAI / LangChain
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.3

# Database
asyncpg==0.29.0
redis==5.0.1

# Utils
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.4
openpyxl==3.1.2